+++ docs/Changelog.md
@@ -14,7 +14,7 @@ Nussknacker versions
     * [#1338](https://github.com/TouK/nussknacker/pull/1338) Defining TypeInformation based on TypingResult
     * [#1343](https://github.com/TouK/nussknacker/pull/1343) Aggregators compute stored types    
     * [#1343](https://github.com/TouK/nussknacker/pull/1359) Improvements in variable output validation 
-
+    * [#1360](https://github.com/TouK/nussknacker/pull/1360) Service query can use global variables
 
 0.3.1 (not released yet) 
 ------------------------